refactor(lib): address review feedback on bootstrap module
- Rename migrations.py -> bootstrap.py to avoid collision with
  db/migrations/ (Alembic scripts).
- Use Path.mkdir() throughout ensure_app_dir for API consistency.
- Comment that metadata=UUIDAuditBase.metadata is intentionally inert
  given create_all=False and Alembic's file-based scripts.
- Use pathlib for the sqlite connection string and config writes.
- Add test_ensure_db_preserves_existing_data: insert a row between
  ensure_db calls and verify it survives.
